// src/types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          extensions?: Json
          operationName?: string
          query?: string
          variables?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      albums_of_week: {
        Row: {
          artist: string | null
          cover_url: string | null
          created_at: string
          id: string
          spotify_id: string
          spotify_url: string | null
          title: string
          week_start_date: string
        }
        Insert: {
          artist?: string | null
          cover_url?: string | null
          created_at?: string
          id?: string
          spotify_id: string
          spotify_url?: string | null
          title: string
          week_start_date: string
        }
        Update: {
          artist?: string | null
          cover_url?: string | null
          created_at?: string
          id?: string
          spotify_id?: string
          spotify_url?: string | null
          title?: string
          week_start_date?: string
        }
        Relationships: []
      }
      suggestions: {
        Row: {
          artist: string | null
          cover_url: string | null
          created_at: string
          id: string
          is_active: boolean
          note: string | null
          spotify_id: string
          spotify_url: string | null
          suggested_by: string | null
          title: string
        }
        Insert: {
          artist?: string | null
          cover_url?: string | null
          created_at?: string
          id?: string
          is_active?: boolean
          note?: string | null
          spotify_id: string
          spotify_url?: string | null
          suggested_by?: string | null
          title: string
        }
        Update: {
          artist?: string | null
          cover_url?: string | null
          created_at?: string
          id?: string
          is_active?: boolean
          note?: string | null
          spotify_id?: string
          spotify_url?: string | null
          suggested_by?: string | null
          title?: string
        }
        Relationships: []
      }
      votes: {
        Row: {
          album_week_id: string
          comment: string | null
          created_at: string
          id: string
          rating: number
          voter: string
        }
        Insert: {
          album_week_id: string
          comment?: string | null
          created_at?: string
          id?: string
          rating: number
          voter: string
        }
        Update: {
          album_week_id?: string
          comment?: string | null
          created_at?: string
          id?: string
          rating?: number
          voter?: string
        }
        Relationships: [
          {
            foreignKeyName: "votes_album_week_id_fkey"
            columns: ["album_week_id"]
            isOneToOne: false
            referencedRelation: "albums_of_week"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never